core/models.py
from django.db import models
from django.db.models.fields import SlugField
from django.template import defaultfilters
from stdimage.models import StdImageField

#SIGNALS
from django.db.models import signals
from django.template.defaultfilters import slugify