requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
Django==3.2.9
django-bootstrap4==21.1
django-stdimage==5.3.0
gunicorn==20.1.0
mysql==0.0.3
mysqlclient==2.0.3
Pillow==8.4.0
PyMySQL==1.0.2
pytz==2021.3
soupsieve==2.3
sqlparse==0.4.2
whitenoise==5.3.0
